+++ AWSCore/Service/AWSServiceEnum.h
@@ -163,6 +163,10 @@ typedef NS_ENUM(NSInteger, AWSRegionType) {
      *  Israel (Tel Aviv)
      */
     AWSRegionILCentral1 NS_SWIFT_NAME(ILCentral1),
+    /**
+     *  Mexico (Central)
+     */
+    AWSRegionMXCentral1 NS_SWIFT_NAME(MXCentral1),
 };
 
 /**