+++ AWSS3/AWSS3Resources.m
@@ -1390,6 +1390,7 @@
         \"eu-west-1\",\
         \"eu-west-2\",\
         \"eu-west-3\",\
+        \"mx-central-1\",\
         \"il-central-1\",\
         \"me-central-1\",\
         \"me-south-1\",\